pyproject.toml
[project]
name = "anahit-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.3",
    "django-ninja>=1.4.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "freezegun>=1.5.2",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
]
